+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- Fixed a bug when working with Arenadata DB 7.4.0 and higher.
+
 ## [14.4.0] - 2026-05-05
 
 ### Changed